+++ CHANGELOG.md
@@ -12,6 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [0.0.11] - 2024-23-09
+### Added
+
+- [Conforme Issue 24](https://github.com/jtonynet/js-med-planner/issues/26):
+  - TDD rota de deletar do recurso `appointment`
+  - Atualizado Swagger
+
 ## [0.0.10] - 2024-23-09
 ### Added
 
@@ -109,6 +116,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Iniciado o [Projeto](https://github.com/users/jtonynet/projects/3) com o commit inicial. Documentação base: Readme Rico, [Diagramas Mermaid](https://github.com/jtonynet/go-products-api/tree/main#diagrams), ADRs: [0001: Registro de Decisões de Arquitetura (ADR)](./docs/architecture/decisions/registro-de-decisoes-de-arquitetura.md) e [0002: Express, Sequelize e Postgress em API TDD Jest com arquitetura de Duas camadas](./docs/architecture/decisions/0002-express-sequelize-e-postgress-tdd-jest-com-arquitetura-de-api-em-duas-camadas.md).
 - Sabemos o que fazer, graças às definições do arquivo __README.md__. Sabemos como fazer graças aos __ADRs__ e documentações vinculadas. Devemos nos organizar em estrutura __Kanban__, guiados pelo modelo Agile, em nosso __Github Project__, e dar o devido prosseguimento às tarefas.
 
+[0.0.11]: https://github.com/jtonynet/js-med-planner/compare/v0.0.10...v0.0.11
 [0.0.10]: https://github.com/jtonynet/js-med-planner/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/jtonynet/js-med-planner/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/jtonynet/js-med-planner/compare/v0.0.7...v0.0.8